Separate out articulated ICP test utilities.
