+++ drake/perception/estimators/dev/test/simple_cuboid.urdf
